split base.py and restructured into its own dir to simplify
