+++ .flake8
@@ -1,6 +1,8 @@
 [flake8]
 ignore = E741,E226,W503,W504
-exclude = 
+per-file-ignores =
+    */__init__.py:F401,F403
+exclude =
     build/,
     .eggs/,
     benchmarks/data,
